Refactor LoggerService for multiple log writers
Revised the LoggerService to support multiple ILogWriter instances instead of a single one. Extracted LoggerServiceConfig to a separate file to improve code organization. Adjusted methods to iterate over multiple log writers for logging operations.
