+++ LoggerServiceSolution/PvWay.LoggerService.nc6/PvWay.LoggerService.nc6.csproj
@@ -16,12 +16,12 @@
     <PackageTags>LoggerSesrvice</PackageTags>
     <PackageReleaseNotes>Refactoring</PackageReleaseNotes>
     <PackageLicenseExpression>MIT</PackageLicenseExpression>
-    <Version>3.0.0</Version>
+    <Version>3.0.1</Version>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="6.0.0" />
-    <PackageReference Include="PvWay.LoggerService.Abstractions.nc6" Version="3.0.0" />
+    <PackageReference Include="PvWay.LoggerService.Abstractions.nc6" Version="3.0.1" />
   </ItemGroup>
 
   <ItemGroup>
